backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
requests
google-genai
httpx